+++ platform/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.3.0] platform
 ### Releases
 - [NPM v1.3.0] https://www.npmjs.com/package/@codat/platform/v/1.3.0 - platform
+
+## 2023-11-13 01:30:04
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.0] platform
+### Releases
+- [NPM v2.0.0] https://www.npmjs.com/package/@codat/platform/v/2.0.0 - platform